Modal Toggle Button implementation
Separated logic for modal toggle into it's own
button component. Applied css styles to and transitions
to give animation effect to this button when the modal is
open and closed. Styles are in their own css file
and are imported into the component itself.

Added personal profile image and Frontend mentor
logo to the public directory.