Refactor: Added Tailwind breakpoint prefix to hover States
After testing on mobile realise that hover states
and focus states don't make much sense as the user
is on mobile. Seems to make sense to me to only
add those rules when the SPA is being viewed on
lg: 64rem (1024px) viewports.

Also noted that becuase I'm using SVG's for the social
icons they do not change colour as per hover state
when in focus states. Removed these classes as they
were doing nothing.

- Something to investigate in future builds. How
to manipulate SVG classes and styling with a state
like :focus.