Refactor: Conditionally Render InfoModal
While researching how to make a focus trap and
build a modal that can be repurposed across future
projects. Conditionally rendering components became
apparant. This makes more sense as opposed to using
codtionally added or removed classes to the modal.

Rather remove it completely from the DOM. Then when
it is called render it.